refactor: 프로필 설정 리팩토링
