+++ src/components/my/types/settings.ts
@@ -1,0 +1,15 @@
+import type { LucideIcon } from "lucide-react";
+
+export type SettingItem = {
+  icon: LucideIcon;
+  label: string;
+  value: string;
+  color: string;
+  onClick: () => void;
+  danger?: boolean;
+};
+
+export type SettingsGroup = {
+  title: string;
+  items: SettingItem[];
+};
